move prefix.go and protocol.go to types/
