+++ app/protocol/keeper/types.go
@@ -1,10 +1,12 @@
 package keeper
 
-import "github.com/irisnet/irishub/types/common"
+import (
+	sdk "github.com/irisnet/irishub/types"
+)
 
 const AppVersionTag = "app_version"
 
 type UpgradeConfig struct {
 	ProposalID uint64
-	Definition common.ProtocolDefinition
+	Definition sdk.ProtocolDefinition
 }
