+++ cmd/irislcd/main.go
@@ -4,7 +4,7 @@ import (
 	"github.com/irisnet/irishub/app"
 	"github.com/irisnet/irishub/lite"
 	_ "github.com/irisnet/irishub/lite/statik"
-	irisInit "github.com/irisnet/irishub/server/init"
+	sdk "github.com/irisnet/irishub/types"
 	"github.com/irisnet/irishub/version"
 	"github.com/spf13/cobra"
 	"github.com/tendermint/tendermint/libs/cli"
@@ -20,7 +20,7 @@ var (
 
 func main() {
 
-	irisInit.InitBech32Prefix()
+	sdk.InitBech32Prefix()
 
 	cobra.EnableCommandSorting = false
 	cdc := app.MakeCodec()
