+++ modules/upgrade/types.go
@@ -2,13 +2,13 @@ package upgrade
 
 import (
 	protocol "github.com/irisnet/irishub/app/protocol/keeper"
-	"github.com/irisnet/irishub/types/common"
+	sdk "github.com/irisnet/irishub/types"
 )
 
 type AppVersion struct {
 	ProposalID uint64
 	Success    bool
-	Protocol   common.ProtocolDefinition
+	Protocol   sdk.ProtocolDefinition
 }
 
 func NewVersion(upgradeConfig protocol.UpgradeConfig, success bool) AppVersion {
