+++ tools/prometheus/server_test.go
@@ -5,7 +5,7 @@ import (
 	"github.com/go-kit/kit/metrics"
 	"github.com/go-kit/kit/metrics/prometheus"
 	"github.com/irisnet/irishub/app"
-	irisInit "github.com/irisnet/irishub/server/init"
+	sdk "github.com/irisnet/irishub/types"
 	stdprometheus "github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/shirou/gopsutil/cpu"
@@ -19,7 +19,7 @@ import (
 )
 
 func TestMetricsCmd(t *testing.T) {
-	irisInit.InitBech32Prefix()
+	sdk.InitBech32Prefix()
 
 	cdc := app.MakeCodec()
 	comm := MonitorCommand(cdc)
