+++ types/prefix.go
@@ -1,8 +1,4 @@
-package init
-
-import (
-	sdk "github.com/irisnet/irishub/types"
-)
+package types
 
 var (
 	// Bech32PrefixAccAddr defines the Bech32 prefix of an account's address
@@ -20,7 +16,7 @@ var (
 )
 
 func InitBech32Prefix() {
-	config := sdk.GetConfig()
+	config := GetConfig()
 	config.SetBech32PrefixForAccount(Bech32PrefixAccAddr, Bech32PrefixAccPub)
 	config.SetBech32PrefixForValidator(Bech32PrefixValAddr, Bech32PrefixValPub)
 	config.SetBech32PrefixForConsensusNode(Bech32PrefixConsAddr, Bech32PrefixConsPub)
